Factor out FailedRequest
