+++ studio/src/pages/RequestorPage/ResponsePanel/ResponseBody/ResponseBody.tsx
@@ -10,7 +10,6 @@ import { cn, isJson, noop, safeParseJson } from "@/utils";
 import {
   CaretDownIcon,
   CaretRightIcon,
-  LinkBreak2Icon,
   QuestionMarkIcon,
 } from "@radix-ui/react-icons";
 import { useState } from "react";
@@ -19,6 +18,7 @@ import {
   type RequestorActiveResponse,
   isRequestorActiveResponse,
 } from "../../store/types";
+import { FailedRequest } from "./FailedRequest";
 import { ResponseBodyBinary } from "./ResponseBodyBinary";
 import { ResponseBodyText } from "./ResponseBodyText";
 
@@ -133,33 +133,6 @@ function UnknownResponse({
           </span>
         </div>
       </CollapsibleBodyContainer>
-    </div>
-  );
-}
-
-export function FailedRequest({
-  response,
-}: { response?: ProxiedRequestResponse | RequestorActiveResponse }) {
-  // TODO - Show a more friendly error message
-  const failureReason = isRequestorActiveResponse(response)
-    ? null
-    : response?.app_responses?.failureReason;
-  const friendlyMessage =
-    failureReason === "fetch failed" ? "Service unreachable" : null;
-  // const failureDetails = response?.app_responses?.failureDetails;
-  return (
-    <div className="h-full pb-8 sm:pb-20 md:pb-32 flex flex-col items-center justify-center p-4">
-      <div className="flex flex-col items-center justify-center p-4">
-        <LinkBreak2Icon className="h-10 w-10 text-red-200" />
-        <div className="mt-4 text-md text-white text-center">
-          {friendlyMessage
-            ? `Request failed: ${friendlyMessage}`
-            : "Request failed"}
-        </div>
-        <div className="mt-2 text-ms text-gray-400 text-center font-light">
-          Make sure your api is up and has FPX Middleware enabled!
-        </div>
-      </div>
     </div>
   );
 }
