+++ studio/src/pages/RequestorPage/ResponsePanel/ResponseBody/index.ts
@@ -1,2 +1,3 @@
 export * from "./ResponseBody";
 export * from "./ResponseBodyText";
+export * from "./FailedRequest";
